requirements.txt
appnope==0.1.0
decorator==4.0.6
coverage==4.0.3
coveralls==1.1
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
djangorestframework==3.3.2
fake-factory==0.5.3
funcsigs==0.4
gnureadline==6.3.3
gunicorn==19.4.5
httplib2==0.9.2
ipdb==0.8.1
ipython==4.1.1
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mock==1.3.0
oauth2==1.9.0.post1
oauthlib==1.0.3
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.6
Pillow==3.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.1.0
Werkzeug==0.11.4
wheel==0.24.0
